Add `createEmulator` and purging JS
